// src/core/api/generated/spring/schemas/MeetingDTOMeetingType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type MeetingDTOMeetingType = typeof MeetingDTOMeetingType[keyof typeof MeetingDTOMeetingType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingDTOMeetingType = {
  VIRTUAL: 'VIRTUAL',
  IN_PERSON: 'IN_PERSON',
  PHONE_CALL: 'PHONE_CALL',
} as const;