// src/core/api/generated/spring/schemas/MeetingDTOMeetingStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type MeetingDTOMeetingStatus = typeof MeetingDTOMeetingStatus[keyof typeof MeetingDTOMeetingStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingDTOMeetingStatus = {
  SCHEDULED: 'SCHEDULED',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED',
  NO_SHOW: 'NO_SHOW',
} as const;