// src/core/api/generated/spring/schemas/MeetingReminderDTOReminderType.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type MeetingReminderDTOReminderType = typeof MeetingReminderDTOReminderType[keyof typeof MeetingReminderDTOReminderType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingReminderDTOReminderType = {
  EMAIL: 'EMAIL',
  SMS: 'SMS',
  PUSH_NOTIFICATION: 'PUSH_NOTIFICATION',
} as const;